Added multi-file support
the -d option can now be used to batch process a directory. Additionally, there were a few tweaks to the codebase, namely, debug output statements are now controlled by a macro which can be left on or off the command line. Also, while it scans for .gif files, I haven't implemented that yet, so they'll always fail.